refactor(CopyLink): 폴더명 수정
